move types in distinct file
